requirements.txt
requests
discord-py-interactions
python-dotenv